requirements.txt
Flask==3.0.3
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
google-cloud-language==2.13.4  # For sentiment analysis
google-cloud-storage==2.16.0   # For Cloud Storage integration
gunicorn==22.0.0
